removed custom-resource overhead
