+++ .github/workflows/pipeline-deploy-workflow.yml
@@ -4,8 +4,9 @@ env:
   AWS_DEPLOY_ROLE_ARN: arn:aws:iam::637747527212:role/github-actions
   REGION: us-east-1
   ACCOUNT: 637747527212
-  SOURCE_BUCKET: my-serverless-image-handler-images
+  SOURCE_BUCKETS: my-serverless-image-handler-images
   SOLUTION_NAME: my-serverless-image-handler
+  STAGE: image
   VERSION: 1
 
 on:
@@ -47,10 +48,6 @@ jobs:
         working-directory: source/constructs
         run: npm install
       
-      - name: npm install for custom resource
-        working-directory: source/custom-resource
-        run: npm install
-      
       - name: npm install image-handler
         working-directory: source/image-handler
         run: npm install
@@ -70,4 +67,4 @@ jobs:
         run: cdk diff
 
       - name: CDK deploy
-        run: overrideWarningsEnabled=false npx cdk deploy --all --parameters SourceBucketsParameter=${{ env.SOURCE_BUCKET }} --require-approval never
+        run: overrideWarningsEnabled=false npx cdk deploy --all --parameters SourceBucketsParameter=${{ env.SOURCE_BUCKETS }} --require-approval never
