--- source/custom-resource/lib/index.ts
@@ -1,6 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-export * from "./enums";
-export * from "./interfaces";
-export * from "./types";
