--- source/custom-resource/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "custom-resource",
-  "version": "6.2.6",
-  "private": true,
-  "description": "Serverless Image Handler custom resource",
-  "license": "Apache-2.0",
-  "author": {
-    "name": "Amazon Web Services",
-    "url": "https://aws.amazon.com/solutions"
-  },
-  "main": "index.js",
-  "scripts": {
-    "clean": "rm -rf node_modules/ dist/ coverage/",
-    "pretest": "npm run clean && npm ci",
-    "test": "jest --coverage --silent",
-    "bump-version": "npm version $(cat ../../VERSION.txt) --allow-same-version"
-  },
-  "dependencies": {
-    "aws-sdk": "^2.1529.0",
-    "axios": "^1.6.5",
-    "moment": "^2.30.0",
-    "uuid": "^9.0.1"
-  },
-  "devDependencies": {
-    "@types/aws-sdk": "^2.7.0",
-    "@types/jest": "^29.5.6",
-    "@types/node": "^20.10.4",
-    "@types/uuid": "^9.0.6",
-    "jest": "^29.7.0",
-    "ts-jest": "^29.1.1",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.3.3"
-  },
-  "overrides": {
-    "semver": "7.5.4"
-  },
-  "resolutions": {
-    "semver": "7.5.4"
-  }
-}
