--- source/solution-utils/jest.config.js
@@ -1,12 +1,0 @@
-module.exports = {
-  roots: ['<rootDir>/test'],
-  testMatch: ['**/*.spec.ts'],
-  transform: {
-    '^.+\\.tsx?$': 'ts-jest'
-  },
-  coverageReporters: [
-    'text',
-    ['lcov', { 'projectRoot': '../' }]
-  ],
-  setupFiles: ['./test/setJestEnvironmentVariables.ts']
-};
