--- source/solution-utils/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "solution-utils",
-  "version": "6.2.6",
-  "private": true,
-  "description": "Utilities to be used within this solution",
-  "license": "Apache-2.0",
-  "author": {
-    "name": "Amazon Web Services",
-    "url": "https://aws.amazon.com/solutions"
-  },
-  "main": "get-options",
-  "typings": "index",
-  "scripts": {
-    "build": "npm run clean && npm install && npm run build:tsc",
-    "build:tsc": "tsc --project tsconfig.json",
-    "clean": "rm -rf node_modules/ dist/ coverage/",
-    "package": "npm run build && npm prune --production && rsync -avrq ./node_modules ./dist",
-    "pretest": "npm run clean && npm install",
-    "test": "jest --coverage --silent",
-    "bump-version": "npm version $(cat ../../VERSION.txt) --allow-same-version"
-  },
-  "devDependencies": {
-    "@types/jest": "^29.5.5",
-    "@types/node": "^20.10.4",
-    "jest": "^29.7.0",
-    "ts-jest": "^29.1.1",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.3.3"
-  },
-  "overrides": {
-    "semver": "7.5.4"
-  },
-  "resolutions": {
-    "semver": "7.5.4"
-  }
-}
