--- source/solution-utils/test/setJestEnvironmentVariables.ts
@@ -1,6 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-process.env.SOLUTION_ID = "solution-id";
-process.env.SOLUTION_VERSION = "solution-version";
-process.env.LOGGING_LEVEL = "VERBOSE";
