--- source/solution-utils/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es6",
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "outDir": "dist",
-    "emitDecoratorMetadata": true,
-    "esModuleInterop": true,
-    "experimentalDecorators": true,
-    "sourceMap": true,
-    "types": ["node", "@types/jest"]
-  },
-  "include": ["**/*.ts"],
-  "exclude": ["package", "dist", "**/*.map"]
-}
